Make Docker on built hosts run as a rooted daemon
Trying to join a pre-existing Swarm using a rootless Docker
daemon results in:

       Error response from daemon: rpc error: code =
       Unavailable desc...

This is due to Docker running rootlessly, and potential firewall
issues. This fix makes the Docker daemon run as root on all hosts and
disables the firewall; the firewall does not matter for this project
as the cluster is meant to be isolated from the outside world.